v2.1.2 (#20)
* Updated TrakerrHandler
Changed TrakerrHandler code to default to warning severity. Added comments to edify the use of the client.

* Fixes for handler update
Fixes a bunch of small bugs to the trakerr handler, such has giving an eventname to debugs and infos, logging fatals as fatals (The logger returns them as criticals), proper default values for the handler constructor. Also fixes the handler to call super instead of calling an explicit constructor to construct with, to actually use the inherited constructor.

* Update trakerr_io
Updates comment.

* Update trakerr_io.py
Fixes incorrect capitalization in a default value.

* Update README.md
Updates some names in handler, and expands on the changes in the three minute guide in the readme.

* Update README.md
Bolds params in the param list in the 3 minute guide.

* Update README.md
Updates readme example.